+++ spec/dummy/dummy_engine/app/controllers/dummy_engine/application_controller.rb
@@ -1,0 +1,4 @@
+module DummyEngine
+  class ApplicationController < ActionController::API
+  end
+end