+++ spec/dummy/dummy_engine/lib/dummy_engine.rb
@@ -1,0 +1,5 @@
+module DummyEngine
+  class Engine < ::Rails::Engine
+    isolate_namespace DummyEngine
+  end
+end