+++ spec/rails/rfc6570_spec.rb
@@ -6,6 +6,10 @@ ActiveSupport::Inflector.inflections(:en) do |inflect|
   inflect.acronym 'API'
 end
 
+DummyEngine::Engine.routes.draw do
+  get '/action' => 'api#action', as: :action
+end
+
 Dummy::Application.routes.draw do
   get '/' => 'api#index', as: :root
 
@@ -20,6 +24,8 @@ Dummy::Application.routes.draw do
   get '/path/*capture/:title' => 'api#index', as: :test4
   get '/path(/*capture)/:title' => 'api#index', as: :test5
   get '/path(/*capture)(/:title)' => 'api#index', as: :test6
+
+  mount DummyEngine::Engine => '/dummy_engine'
 end
 
 class APIController < ApplicationController
@@ -49,6 +55,18 @@ class APIController < ApplicationController
   end
 end
 
+class DummyEngine::APIController < DummyEngine::ApplicationController
+  rfc6570_params action: %i[param3]
+  def action
+    render json: {
+      ref: action_path,
+      template: action_rfc6570,
+      template_url: action_url_rfc6570,
+      template_path: action_path_rfc6570,
+    }
+  end
+end
+
 describe Rails::RFC6570, type: :request do
   let(:host) { 'http://www.example.com' }
   let(:json) { JSON.parse response.body }
@@ -58,7 +76,7 @@ describe Rails::RFC6570, type: :request do
 
     it 'returns list of all parsed and named routes' do
       expect(json.keys).to match_array \
-        %w[root action test1 test2 test3 test4 test5 test6]
+        %w[root action test1 test2 test3 test4 test5 test6 dummy_engine]
     end
 
     it 'includes known parameters' do
@@ -132,6 +150,26 @@ describe Rails::RFC6570, type: :request do
       it 'prefixes origin script name' do
         expect(json['template']).to eq "#{host}/fuubar/action{?param1,param2}"
       end
+    end
+  end
+
+  context 'action in Rails Engine' do
+    before { get '/dummy_engine/action', headers: headers }
+
+    it 'includes URL helpers' do
+      expect(response).to have_http_status :ok
+    end
+
+    it 'allows to return and render templates' do
+      expect(json['template']).to eq "#{host}/dummy_engine/action{?param3}"
+    end
+
+    it 'allows to return and render url templates' do
+      expect(json['template_url']).to eq "#{host}/dummy_engine/action{?param3}"
+    end
+
+    it 'allows to return and render path templates' do
+      expect(json['template_path']).to eq '/dummy_engine/action{?param3}'
     end
   end
 end